// web/client/models/File.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type File = {
    id: string;
    url?: string;
    /**
     * Имя файла, как оно было на компьютере пользователя
     */
    file_name: string;
};